requirements.txt
playwright==1.55.0
requests
python-dotenv
